requirements.txt
pyzipper
python-mpv
PyQt5
qtpy
qtawesome
